split up hint modal into separate component
